+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.27.0 - 2024-11-27
+
+### Fixed
+
+- Fixed stuck detection with buffer columns
+
+### Added
+
+- Zen mode rewards are now also visible on level track
+
 ## Version 1.26.0 - 2024-11-24
 
 ### Added